backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
google-generativeai==0.3.2
chromadb==0.5.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
asyncpg==0.29.0
Pillow==10.4.0
pytest==8.4.2
pytest-asyncio==0.23.8
requests==2.32.3
